requirements.txt
Flask>=3.0.2
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-Mail>=0.9.1
psycopg2-binary>=2.9.9
qrcode>=7.4.2
Pillow>=10.2.0
python-dotenv>=1.0.1